Move component files from design system into the components folder.
