+++ CLAUDE.md
@@ -20,18 +20,17 @@ CI runs typecheck → lint → test on PRs to main.
 2. `DefaultBudgetIdFetcher` retrieves the user's budget via `YnabService` → stored in `BudgetProvider` context
 3. Components fetch accounts/transactions through `YnabService` (constructed with bearer token)
 4. Pure calculation functions in `calculation-functions/` compute max balances and running balances
-5. Results rendered through components using the `design-system/` component library
+5. Results rendered through components in `components/`
 
 ### Key Directories
 
 - **`app/`** — Next.js App Router pages (home, account detail, help, settings, privacy policy)
-- **`components/`** — Application-level React components (most are `"use client"`)
+- **`components/`** — Application-level React components and UI primitives (Button, Table, Typography, etc.)
 - **`contexts/`** — React Context + useReducer for state (`budgetContext`, `accountsContext`)
 - **`services/ynab-service/`** — `YnabService` class wrapping the `ynab` SDK
 - **`services/tokenManager.ts`** — OAuth token persistence in localStorage
 - **`calculation-functions/`** — Pure functions: `getMaxBalances`, `getTransactionsWithBalances`
 - **`utility-functions/`** — Data fetching helpers (`getAccounts`, `getTransactionsForAccount`, `getDefaultBudgetId`)
-- **`design-system/`** — Reusable UI primitives (Button, Table, Typography, etc.)
 - **`formatters/`** — Display formatting (e.g., `formatAmount` for currency)
 - **`types/`** — TypeScript type definitions (`Account`, `TransactionWithBalance`, `MaxBalanceTransaction`)
 - **`hooks/`** — Custom hooks (`useYnabOauthToken`, `useLocalStorage`)
