+++ app/account/[id]/AccountDetailPageHeading.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import { useAccount } from "@/contexts/accountsContext";
-import Heading1 from "@/design-system/headings/heading1/Heading1";
+import Heading1 from "@/components/headings/heading1/Heading1";
 
 export default function AccountDetailPageHeading({
   accountId,
