+++ app/help/page.tsx
@@ -1,5 +1,5 @@
-import Heading1 from "@/design-system/headings/heading1/Heading1";
-import Heading2 from "@/design-system/headings/heading2/Heading2";
+import Heading1 from "@/components/headings/heading1/Heading1";
+import Heading2 from "@/components/headings/heading2/Heading2";
 
 export default function HelpPage() {
   return (
