+++ app/privacy-policy/page.tsx
@@ -1,6 +1,6 @@
-import Heading1 from "@/design-system/headings/heading1/Heading1";
-import Heading2 from "@/design-system/headings/heading2/Heading2";
-import Paragraph from "@/design-system/paragraph/Paragraph";
+import Heading1 from "@/components/headings/heading1/Heading1";
+import Heading2 from "@/components/headings/heading2/Heading2";
+import Paragraph from "@/components/paragraph/Paragraph";
 
 export default function PrivacyPolicy() {
   return (
