+++ components/Accounts.tsx
@@ -1,4 +1,4 @@
-import Heading1 from "@/design-system/headings/heading1/Heading1";
+import Heading1 from "@/components/headings/heading1/Heading1";
 import AccountTable from "./AccountTable";
 
 export default function Accounts() {
