+++ components/headings/heading2/Heading2.tsx
