+++ components/table/enums.ts
